PRACMIG-512: Move the e2e test lambda here
The lambda has also been updated to assume that the source location of
 the DataSync Task is an NFS file share, rather than an S3 bucket.

The existing lambda, that assumes the source location is an S3 bucket
 has been left in place for now.

As part of creating a second lambda within the the create datasync task project, the agent activator lambda is now built into its own sub-directory of `build/`.

For some reason, the 2 extra libraries that it requires (`polling` and `pyNfsClient`) don't seem to like being installed using pipenv - when I try and explicitly install the latest version (according to their pypi pages) it says those versions don't exist, and when I try and just install them without specifying a version, it doesn't fail, but it also doesn't seem to install them into the .venv directory (and so VSCode doesn't recognise the imports in the handler file).